+++ dataworkspace/dataworkspace/apps/api_v1/views.py
@@ -148,10 +148,7 @@ def application_api_PUT(request, public_host):
         )
 
     try:
-        (
-            application_template,
-            public_host_data,
-        ) = application_template_and_data_from_host(public_host)
+        application_template, _ = application_template_and_data_from_host(public_host)
     except ApplicationTemplate.DoesNotExist:
         return JsonResponse(
             {'message': 'Application template does not exist'}, status=400
@@ -211,7 +208,6 @@ def application_api_PUT(request, public_host):
         application_template.spawner,
         request.user.email,
         str(request.user.profile.sso_id),
-        public_host_data,
         application_instance.id,
         application_template.spawner_options,
         credentials,